cs40l26: Fix the composed effect delay problem
When user typed the gboard back-to-back, HAL will send a STOP call right
after it sends a START call. At this moment, driver is under an
activating process and the vibe_state is not 1, so the erasing process
will be forced to wait for the FW's completed signal.
This makes STOP call consumes lots of waiting time and user would still
feels a delay gap under back-to-back typing behavior.

This patch removes the erased process from the STOP call and let it be
done in the completed process.

Bug: 263476381
Test: atest PtsVibratorHalTestSuite \
            PtsHapticsTestCases \
            PtsHapticsFeatureTestCases \
            VibratorHalCs40l26TestSuite \
            VtsHalVibratorManagerTargetTest \
            VtsHalVibratorTargetTest \
            android.os.cts.VibratorTest \
            android.os.cts.VibratorManagerTest \
            android.os.cts.VibrationEffectTest \
            android.os.cts.VibrationAttributesTest \
            android.os.cts.CombinedVibrationTest
Test: Type Gboard with five fingers

Change-Id: I02dfdae4e9e9334a253106f631c933228398b077