powerhint.json: Update uclamp_min hint to account for margin
Upstream has fixed relationship between uclamp_min and margin which
means uclamp_min value between 80-100% of capacity of the CPU is a valid
value and will not cause upmigration.

To keep the current hints in powerhint.json the same, we must update
uclamp_min to be capacity_orig_of(medium_cpu) + 1.

This change is dependent on uclamp vs fits margin in GKI, vendor
modules. And an improvement in sched_pixel to ignore the 25% boost if
the only reason to request the frequency is uclamp_min.

To cater for the later - we multiply each uclamp_min value with 1.25
- except ones that look intentionally set around capacity boundaries.

Bug: 261695590
Test: PtsUiBench
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I60f4d7cd2cecf7ff4d73fa0cee1a09522b1d9bfd