Increase keyguardDrawnTimeout to reduce fold->AOD flickers
This is a temporary measure to decrease the chances of flickers when folding to aod, while we work in parallel to decrease latency.

According to ww metrics, Fold to AoD takes:

p50: 714ms
p89: 1020ms
p95: 1147ms
p99: 1577ms

Once we hit the keyguard drawn timeout, the screen is turned on . However, in some cases sysui and launcher are not ready, and this results with the screen on while showing the unfolded layout, with a rapid transition to the folded one (according to pitot data[1], this happens in 11% of the fold->aod).

This cl sets the timeout to 1700ms (slightly higher than p99). In this way, the chances to have flicker are greatly reduced, but the underlying latency problem remains.
In parallel, we're looking to decrease this latency (see bug).

[1] https://screenshot.googleplex.com/9gV9nEfpstkJmCt

Bug: 243015678
Test: fold->aod many times, no flicker.
Change-Id: I0738901279136dd7577016a3ba46dc4c5b9a86f1