add ALLOW_MISSING_DEPENDENCIES := true
This is currently needed for inclusion of felix-services system_server JAR.

comet has an equivalent comet-services JAR, but ALLOW_MISSING_DEPENDENCIES is already set for it in
device/google/zumapro.

It's not clear whether stock Pixel OS uses ALLOW_MISSING_DEPENDENCIES for felix too or if it uses
some other configuration.